serverb/soap/requirements.txt
spyne
lxml
pymongo
pika
python-jose[cryptography]
requests